history AI Generate Plan

+++ Models/ItineraryResponse.cs
@@ -18,7 +18,7 @@ namespace TripWiseAPI.Model
 
         public List<ItineraryDay> Itinerary { get; set; }
         public int TotalEstimatedCost { get; set; }
-        public int Budget { get; set; }
+        public Decimal Budget { get; set; }
 
         public string SuggestedAccommodation { get; set; }
